[+]전체적인 레이아웃 보완
