+++ client/app/src/main/java/yuhan/hgcq/client/view/Modify.java
@@ -1,18 +1,8 @@
 package yuhan.hgcq.client.view;
 
-import android.content.Context;
-import android.content.DialogInterface;
-import android.graphics.Rect;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.Looper;
-import android.view.MotionEvent;
-import android.view.View;
-import android.view.inputmethod.InputMethodManager;
-import android.widget.EditText;
 
 import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
@@ -22,71 +12,17 @@ import yuhan.hgcq.client.R;
 
 public class Modify extends AppCompatActivity {
 
-    /* View */
-
-    /* 서버와 통신 */
-
-    /* Toast */
-    Handler handler = new Handler(Looper.getMainLooper());
-
-    /* Request Code */
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        getSupportActionBar().setTitle("나의 정보");
+        getSupportActionBar().setTitle("정보 수정");
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         super.onCreate(savedInstanceState);
-
         EdgeToEdge.enable(this);
         setContentView(R.layout.activity_modify);
-
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
-
-        /* 서버와 연결할 Controller 생성 */
-
-        /* View와 Layout 연결 */
-
-        /* 관련된 페이지 */
-
-        /* 받아 올 값 */
-
-        /* 공유 초기 설정 */
-    }
-
-    /* Confirm 창 */
-    public void onClick_setting_costume_cancel(String message,
-                                               DialogInterface.OnClickListener positive,
-                                               DialogInterface.OnClickListener negative) {
-
-        new AlertDialog.Builder(this)
-                .setTitle("Recoder")
-                .setMessage(message)
-                .setIcon(R.drawable.album)
-                .setPositiveButton(android.R.string.yes, positive)
-                .setNegativeButton(android.R.string.no, negative)
-                .show();
-    }
-
-    @Override
-    public boolean dispatchTouchEvent(MotionEvent ev) {
-        if (ev.getAction() == MotionEvent.ACTION_DOWN) {
-            View v = getCurrentFocus();
-            if (v instanceof EditText) {
-                Rect outRect = new Rect();
-                v.getGlobalVisibleRect(outRect);
-                if (!outRect.contains((int) ev.getRawX(), (int) ev.getRawY())) {
-                    v.clearFocus();
-                    InputMethodManager imm = (InputMethodManager) getSystemService(Context.INPUT_METHOD_SERVICE);
-                    if (imm != null) {
-                        imm.hideSoftInputFromWindow(v.getWindowToken(), 0);
-                    }
-                }
-            }
-        }
-        return super.dispatchTouchEvent(ev);
     }
 }
