+++ client/app/src/main/res/drawable/border.xml
@@ -2,6 +2,6 @@
 
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle" android:thickness="0dp">
-    <stroke android:width="1dp" android:color="#AAAAAA"/>
+    <stroke android:width="0dp" android:color="#AAAAAA"/>
     <solid android:color="#ffffff"/>
 </shape>
