+++ client/app/src/main/res/drawable/chattext.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:padding="10dp"
+    android:shape="rectangle">
+
+    <stroke android:width="2dp"
+        android:color="#E2E2E2"/>
+
+    <solid android:color="#FFFFFF"/>
+    <corners
+        android:topLeftRadius="6dp"
+        android:topRightRadius="6dp"
+        android:bottomLeftRadius="6dp"
+        android:bottomRightRadius="6dp"/>
+
+</shape>
