+++ client/app/src/main/res/layout/activity_album_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     tools:context=".view.AlbumMain">
 
-   <FrameLayout
+  <FrameLayout
         android:id="@+id/main_container"
         android:layout_width="match_parent"
         android:layout_height="0dp"
@@ -24,26 +24,25 @@
         android:background="#D3D3D3"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.196" />
+        app:layout_constraintVertical_bias="0.152" />
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/albumList"
-        android:layout_width="366dp"
-        android:layout_height="472dp"
-        android:layout_marginEnd="20dp"
-        android:gravity="center"
-        android:padding="17dp"
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.629"
-        app:spanCount="2" />
+  <androidx.recyclerview.widget.RecyclerView
+      android:id="@+id/albumList"
+      android:layout_width="390dp"
+      android:layout_height="475dp"
+      android:layout_marginEnd="8dp"
+      android:gravity="center"
+      android:padding="17dp"
+      app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintVertical_bias="0.435"
+      app:spanCount="2" />
 
-    <TextView
+  <TextView
         android:id="@+id/empty"
         android:layout_width="200dp"
         android:layout_height="50dp"
@@ -51,7 +50,7 @@
         android:gravity="center"
         android:text="비어있음"
         android:textSize="18dp"
-        android:fontFamily="@font/notosans_bold"
+        android:fontFamily="@font/notosans_medium"
         android:visibility="invisible"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.497"
@@ -84,21 +83,20 @@
         app:layout_constraintTop_toTopOf="parent"
         app:tint="#234567" />
 
-    <ImageButton
-        android:id="@+id/albumplus"
-        android:layout_width="35dp"
-        android:layout_height="29dp"
-        android:layout_marginTop="80dp"
-        android:background="#00000000"
-        android:elevation="8dp"
-        android:scaleType="fitCenter"
-        android:src="@drawable/plusicon"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.787"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+  <ImageButton
+      android:id="@+id/albumplus"
+      android:layout_width="50dp"
+      android:layout_height="45dp"
+      android:layout_marginTop="592dp"
+      android:background="#00000000"
+      android:scaleType="fitCenter"
+      android:src="@drawable/plusicon"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintHorizontal_bias="0.498"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toTopOf="parent" />
 
-    <ImageButton
+  <ImageButton
         android:id="@+id/auto"
         android:layout_width="35dp"
         android:layout_height="29dp"
@@ -123,19 +121,6 @@
         app:layout_constraintHorizontal_bias="0.983"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
-
- <ImageButton
-     android:id="@+id/photoTrash"
-     android:layout_width="39dp"
-     android:layout_height="31dp"
-     android:layout_marginTop="80dp"
-     android:background="#00000000"
-     android:scaleType="fitCenter"
-     android:src="@drawable/phototrash"
-     app:layout_constraintEnd_toEndOf="parent"
-     app:layout_constraintHorizontal_bias="0.69"
-     app:layout_constraintStart_toStartOf="parent"
-     app:layout_constraintTop_toTopOf="parent" />
 
  <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottom_navigation_view"
