+++ client/app/src/main/res/layout/activity_chat.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     tools:context=".view.AlbumMain">
 
-    <FrameLayout
+   <FrameLayout
         android:id="@+id/main_container"
         android:layout_width="match_parent"
         android:layout_height="0dp"
@@ -54,46 +54,50 @@
         app:layout_constraintVertical_bias="0.906"
         app:tint="#234567" />
 
-    <ImageView
-        android:id="@+id/profile"
-        android:layout_width="41dp"
-        android:layout_height="41dp"
-        android:layout_marginTop="568dp"
-        android:src="@drawable/profile"
-        android:text="팔로잉"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.030"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+ <EditText
+     android:id="@+id/chatting"
+     android:layout_width="352dp"
+     android:layout_height="44dp"
+     android:layout_marginTop="596dp"
+     android:background="@drawable/edittext"
+     android:fontFamily="@font/notosans_medium"
+     android:gravity="center"
+     android:hint="채팅을 입력해주세요"
+     android:textSize="18dp"
+     app:layout_constraintEnd_toEndOf="parent"
+     app:layout_constraintHorizontal_bias="0.152"
+     app:layout_constraintStart_toStartOf="parent"
+     app:layout_constraintTop_toTopOf="parent" />
 
-    <EditText
-        android:id="@+id/chatting"
-        android:layout_width="271dp"
-        android:layout_height="42dp"
-        android:layout_marginTop="568dp"
-        android:background="@drawable/edittext"
-        android:fontFamily="@font/notosans_medium"
-        android:gravity="center"
-        android:hint="채팅을 입력해주세요"
-        android:textSize="20dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+ <TextView
+     android:id="@+id/AlbumText"
+     android:layout_width="350dp"
+     android:layout_height="47dp"
+     android:layout_marginTop="28dp"
+     android:background="@drawable/text"
+     android:fontFamily="@font/notosans_medium"
+     android:gravity="center"
+     android:text="Title"
+     android:textSize="17dp"
+     app:layout_constraintEnd_toEndOf="parent"
+     app:layout_constraintHorizontal_bias="0.491"
+     app:layout_constraintStart_toStartOf="parent"
+     app:layout_constraintTop_toTopOf="parent" />
 
-    <ImageButton
-        android:id="@+id/arrow"
-        android:layout_width="35dp"
-        android:layout_height="35dp"
-        android:background="#00000000"
-        android:scaleType="fitCenter"
-        android:src="@drawable/arrow"
-        android:onClick="onClick_setting_costume_save"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.960"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.86" />
+ <ImageButton
+     android:id="@+id/arrow"
+     android:layout_width="35dp"
+     android:layout_height="35dp"
+     android:background="#00000000"
+     android:onClick="onClick_setting_costume_save"
+     android:scaleType="fitCenter"
+     android:src="@drawable/arrow"
+     app:layout_constraintBottom_toBottomOf="parent"
+     app:layout_constraintEnd_toEndOf="parent"
+     app:layout_constraintHorizontal_bias="0.969"
+     app:layout_constraintStart_toStartOf="parent"
+     app:layout_constraintTop_toTopOf="parent"
+     app:layout_constraintVertical_bias="0.860" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
