+++ client/app/src/main/res/layout/activity_gallery.xml
@@ -8,16 +8,47 @@
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/photoList"
-        android:layout_width="412dp"
-        android:layout_height="520dp"
+        android:layout_width="390dp"
+        android:layout_height="475dp"
+        android:layout_marginEnd="4dp"
         android:gravity="center"
         android:padding="17dp"
         app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.6"
+        app:layout_constraintVertical_bias="0.375"
         app:spanCount="4" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/albumList"
+        android:layout_width="350dp"
+        android:layout_height="410dp"
+        android:layout_marginEnd="32dp"
+        android:background="@drawable/text"
+        android:gravity="center"
+        android:padding="17dp"
+        android:visibility="invisible"
+        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.408"
+        app:spanCount="2" />
+
+    <TextView
+        android:id="@+id/date"
+        android:layout_width="200dp"
+        android:layout_height="50dp"
+        android:layout_marginTop="28dp"
+        android:fontFamily="@font/notosans_bold"
+        android:gravity="center"
+        android:text="date"
+        android:textSize="18dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/empty"
@@ -27,7 +58,7 @@
         android:gravity="center"
         android:text="비어있음"
         android:textSize="18dp"
-        android:fontFamily="@font/notosans_bold"
+        android:fontFamily="@font/notosans_medium"
         android:visibility="invisible"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.497"
@@ -38,75 +69,14 @@
         android:id="@+id/move"
         android:layout_width="33dp"
         android:layout_height="22dp"
-        android:layout_marginTop="104dp"
+        android:layout_marginTop="56dp"
         android:background="#00000000"
         android:scaleType="fitCenter"
         android:src="@drawable/move"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.87"
+        app:layout_constraintHorizontal_bias="0.989"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageButton
-        android:id="@+id/photoPlus"
-        android:layout_width="33dp"
-        android:layout_height="22dp"
-        android:layout_marginTop="104dp"
-        android:background="#00000000"
-        android:scaleType="fitCenter"
-        android:src="@drawable/plusicon"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.957"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"/>
-    <ImageButton
-        android:id="@+id/albumDelete"
-        android:layout_width="35dp"
-        android:layout_height="25dp"
-        android:layout_marginTop="104dp"
-        android:background="#00000000"
-        android:scaleType="fitCenter"
-        android:onClick=""
-        android:src="@drawable/albumdelete"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.590"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"/>
-    <ImageButton
-        android:id="@+id/photoTrash"
-        android:layout_width="39dp"
-        android:layout_height="31dp"
-        android:layout_marginTop="96dp"
-        android:background="#00000000"
-        android:scaleType="fitCenter"
-        android:src="@drawable/phototrash"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.69"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageButton
-        android:id="@+id/chat"
-        android:layout_width="33dp"
-        android:layout_height="22dp"
-        android:layout_marginTop="104dp"
-        android:background="#00000000"
-        android:scaleType="fitCenter"
-        android:src="@drawable/chat"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.783"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-    <View
-        android:layout_width="417dp"
-        android:layout_height="2dp"
-        android:background="#D3D3D3"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.18" />
 
     <FrameLayout
         android:id="@+id/main_container"
@@ -117,7 +87,57 @@
         app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0" />
+        app:layout_constraintVertical_bias="0.0" />
+
+    <ImageButton
+        android:id="@+id/photoPlus"
+        android:layout_width="50dp"
+        android:layout_height="45dp"
+        android:layout_marginTop="584dp"
+        android:background="#00000000"
+        android:scaleType="fitCenter"
+        android:src="@drawable/plusicon"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageButton
+        android:id="@+id/photoTrash"
+        android:layout_width="39dp"
+        android:layout_height="31dp"
+        android:layout_marginTop="48dp"
+        android:background="#00000000"
+        android:scaleType="fitCenter"
+        android:src="@drawable/phototrash"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.9"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageButton
+        android:id="@+id/chat"
+        android:layout_width="33dp"
+        android:layout_height="22dp"
+        android:layout_marginTop="56dp"
+        android:background="#00000000"
+        android:scaleType="fitCenter"
+        android:src="@drawable/chat"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.044"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <View
+        android:layout_width="417dp"
+        android:layout_height="2dp"
+        android:background="#D3D3D3"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.111" />
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottom_navigation_view"
