+++ client/app/src/main/res/layout/activity_like.xml
@@ -17,7 +17,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.497"
-        app:spanCount="5" />
+        app:spanCount="2" />
 
     <TextView
         android:id="@+id/empty"
@@ -27,7 +27,7 @@
         android:gravity="center"
         android:text="비어있음"
         android:textSize="18dp"
-        android:fontFamily="@font/notosans_bold"
+        android:fontFamily="@font/notosans_medium"
         android:visibility="invisible"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.497"
