+++ client/app/src/main/res/layout/item_album.xml
@@ -5,12 +5,14 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    android:padding="16dp"
+    android:padding="0dp"
     android:id="@+id/main">
 
     <ImageView
-        android:layout_width="186dp"
-        android:layout_height="138dp"
+        android:id="@+id/basicAlbumImage"
+        android:layout_width="160dp"
+        android:layout_height="130dp"
+        android:background="@drawable/basicalbum"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -18,49 +20,34 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0" />
 
-    <TextView
-        android:id="@+id/date"
-        android:layout_width="70dp"
-        android:layout_height="40dp"
-        android:fontFamily="@font/notosans_medium"
-        android:gravity="center"
-        android:text="date"
-        android:textSize="15sp"
-        app:layout_constraintBottom_toBottomOf="parent"
+    <ImageButton
+        android:id="@+id/albumDelete"
+        android:layout_width="20dp"
+        android:layout_height="20dp"
+        android:layout_marginTop="108dp"
+        android:background="#00000000"
+        android:scaleType="fitCenter"
+        android:src="@drawable/albumdelete"
+        android:visibility="invisible"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintHorizontal_bias="0.409"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.27" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/title"
-        android:layout_width="90dp"
-        android:layout_height="40dp"
-        android:fontFamily="@font/notosans_medium"
-        android:gravity="center"
-        android:text="title"
-        android:textSize="18sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.206" />
-
-    <TextView
-        android:id="@+id/createGroupText"
         android:layout_width="70dp"
         android:layout_height="40dp"
         android:fontFamily="@font/notosans_medium"
         android:gravity="center"
-        android:text="Groupname"
+        android:padding="0dp"
+        android:text="title"
         android:textSize="15sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.145" />
+        app:layout_constraintVertical_bias="0.185" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
