+++ client/app/src/main/res/layout/item_album_trash.xml
@@ -5,12 +5,12 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    android:padding="16dp"
+    android:padding="0dp"
     android:id="@+id/main">
 
     <ImageView
-        android:layout_width="186dp"
-        android:layout_height="138dp"
+        android:layout_width="160dp"
+        android:layout_height="130dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -20,32 +20,34 @@
 
     <TextView
         android:id="@+id/date"
+        android:layout_width="75dp"
+        android:layout_height="31dp"
+        android:fontFamily="@font/notosans_medium"
+        android:gravity="center"
+        android:padding="0dp"
+        android:text="date"
+        android:textSize="13sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.242" />
+
+    <TextView
+        android:id="@+id/title"
         android:layout_width="70dp"
         android:layout_height="40dp"
         android:fontFamily="@font/notosans_medium"
         android:gravity="center"
-        android:text="date"
+        android:padding="0dp"
+        android:text="title"
         android:textSize="15sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.27" />
-
-    <TextView
-        android:id="@+id/title"
-        android:layout_width="90dp"
-        android:layout_height="40dp"
-        android:fontFamily="@font/notosans_medium"
-        android:gravity="center"
-        android:text="title"
-        android:textSize="18sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.206" />
+        app:layout_constraintVertical_bias="0.188" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
