+++ client/app/src/main/res/layout/item_chat.xml
@@ -5,14 +5,14 @@
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="horizontal"
-    android:padding="16dp"
+    android:padding="0dp"
     android:id="@+id/main">
 
     <ImageView
         android:id="@+id/profile"
-        android:layout_width="66dp"
-        android:layout_height="54dp"
-        android:layout_marginTop="24dp"
+        android:layout_width="60dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="12dp"
         android:src="@drawable/profile"
         android:text="팔로잉"
         app:layout_constraintStart_toStartOf="parent"
@@ -20,28 +20,28 @@
 
     <TextView
         android:id="@+id/name"
-        android:layout_width="130dp"
-        android:layout_height="44dp"
-        android:layout_marginTop="4dp"
+        android:layout_width="80dp"
+        android:layout_height="30dp"
         android:fontFamily="@font/notosans_regular"
+        android:padding="0dp"
         android:text="name"
-        android:textSize="20dp"
+        android:textSize="15dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.289"
+        app:layout_constraintHorizontal_bias="0.181"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/chatting"
-        android:layout_width="289dp"
-        android:layout_height="41dp"
-        android:layout_marginTop="48dp"
-        android:text="chat"
-        android:textSize="20dp"
+        android:layout_width="290dp"
+        android:layout_height="35dp"
+        android:layout_marginTop="28dp"
+        android:background="@drawable/text"
         android:fontFamily="@font/notosans_medium"
+        android:padding="0dp"
+        android:text="chat"
+        android:textSize="18dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.8"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
