+++ client/app/src/main/res/layout/item_group.xml
@@ -5,12 +5,14 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    android:padding="16dp"
+    android:padding="0dp"
     android:id="@+id/main">
 
     <ImageView
-        android:layout_width="186dp"
-        android:layout_height="138dp"
+        android:id="@+id/basicGroupImage"
+        android:background="@drawable/basicgroup"
+        android:layout_width="160dp"
+        android:layout_height="130dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -18,48 +20,35 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0" />
 
-    <TextView
-        android:id="@+id/date"
-        android:layout_width="70dp"
-        android:layout_height="40dp"
-        android:fontFamily="@font/notosans_medium"
-        android:gravity="center"
-        android:text="date"
-        android:textSize="15sp"
-        app:layout_constraintBottom_toBottomOf="parent"
+    <ImageButton
+        android:id="@+id/groupset"
+        android:layout_width="33dp"
+        android:layout_height="34dp"
+        android:layout_marginTop="96dp"
+        android:background="#00000000"
+        android:padding="0dp"
+        android:scaleType="fitCenter"
+        android:src="@drawable/groupset"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintHorizontal_bias="0.335"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.27" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/title"
+        android:id="@+id/groupText"
         android:layout_width="90dp"
         android:layout_height="40dp"
         android:fontFamily="@font/notosans_medium"
         android:gravity="center"
-        android:text="title"
+        android:padding="0dp"
+        android:text="group"
         android:textSize="18sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.206" />
-
-    <ImageButton
-        android:id="@+id/groupset"
-        android:layout_width="33dp"
-        android:layout_height="34dp"
-        android:layout_marginTop="104dp"
-        android:background="#00000000"
-        android:scaleType="fitCenter"
-        android:src="@drawable/groupset"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.442"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"/>
+        app:layout_constraintVertical_bias="0.188" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
