+++ client/app/src/main/res/layout/item_group_setting.xml
@@ -5,13 +5,12 @@
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="horizontal"
-    android:padding="16dp">
+    android:padding="0dp">
 
     <ImageView
         android:id="@+id/profile"
-        android:layout_width="66dp"
-        android:layout_height="54dp"
-        android:layout_marginTop="24dp"
+        android:layout_width="60dp"
+        android:layout_height="40dp"
         android:src="@drawable/profile"
         android:text="팔로잉"
         app:layout_constraintStart_toStartOf="parent"
@@ -19,12 +18,12 @@
 
     <Button
         android:id="@+id/delete"
-        android:layout_width="89dp"
-        android:layout_height="59dp"
-        android:layout_marginTop="16dp"
-        android:text="삭제"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
         android:fontFamily="@font/notosans_bold"
         android:onClick="onClick_setting_costume_save"
+        android:padding="0dp"
+        android:text="삭제"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
@@ -32,27 +31,14 @@
 
     <TextView
         android:id="@+id/name"
-        android:layout_width="128dp"
-        android:layout_height="45dp"
-        android:layout_marginTop="8dp"
+        android:layout_width="80dp"
+        android:layout_height="30dp"
         android:fontFamily="@font/notosans_regular"
+        android:padding="0dp"
         android:text="name"
-        android:textSize="20dp"
+        android:textSize="15dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.29"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/email"
-        android:layout_width="159dp"
-        android:layout_height="43dp"
-        android:layout_marginTop="56dp"
-        android:fontFamily="@font/notosans_regular"
-        android:text="email"
-        android:textSize="20dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.331"
+        app:layout_constraintHorizontal_bias="0.2"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
