+++ client/app/src/main/res/layout/item_photo_date.xml
@@ -7,27 +7,17 @@
     android:orientation="vertical"
     android:padding="0dp">
 
-    <ImageView
-        android:id="@+id/photo"
-        android:layout_width="85dp"
-        android:layout_height="85dp"
-        android:layout_marginTop="40dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
     <TextView
         android:id="@+id/date"
-        android:layout_width="115dp"
-        android:layout_height="40dp"
+        android:layout_width="200dp"
+        android:layout_height="50dp"
+        android:fontFamily="@font/notosans_bold"
         android:gravity="center"
+        android:padding="0dp"
         android:text="date"
         android:textSize="20sp"
-        android:padding="0dp"
-        android:fontFamily="@font/notosans_medium"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
