+++ client/app/src/main/res/layout/item_photo_image.xml
@@ -4,29 +4,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:padding="0dp">
+    android:orientation="vertical">
 
     <ImageView
         android:id="@+id/photo"
-        android:layout_width="85dp"
-        android:layout_height="85dp"
-        android:layout_marginTop="40dp"
+        android:layout_width="90dp"
+        android:layout_height="90dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <TextView
-        android:id="@+id/date"
-        android:layout_width="115dp"
-        android:layout_height="40dp"
-        android:fontFamily="@font/notosans_medium"
-        android:gravity="center"
-        android:padding="0dp"
-        android:text="date"
-        android:textSize="20sp"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
