+++ client/app/src/main/res/layout/item_photo_trash.xml
@@ -5,12 +5,12 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    android:padding="16dp">
-
+    android:padding="0dp">
     <ImageView
-        android:layout_width="93dp"
-        android:layout_height="90dp"
-        android:layout_marginTop="68dp"
+        android:id="@+id/photo"
+        android:layout_width="85dp"
+        android:layout_height="85dp"
+        android:layout_marginTop="40dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
@@ -18,14 +18,13 @@
 
     <TextView
         android:id="@+id/date"
-        android:layout_width="130dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="8dp"
-        android:fontFamily="@font/notosans_bold"
+        android:layout_width="115dp"
+        android:layout_height="40dp"
+        android:fontFamily="@font/notosans_medium"
         android:gravity="center"
+        android:padding="0dp"
         android:text="date"
-        android:textSize="25sp"
-        app:layout_constraintEnd_toEndOf="parent"
+        android:textSize="20sp"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
